06-streaming/homework/pyproject.toml
[tool.poetry]
name = "homework-kafka"
version = "0.1.0"
description = ""
authors = ["Lily Chau <chauyeeki@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
kafka-python = "^2.0.2"
pandas = "^2.2.1"
jupyter = "^1.0.0"
pyspark = "^3.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
